fix error handle

+++ src/entities/subscriptions.js
@@ -1,6 +1,13 @@
+import * as User from "koa";
+
 const Subscription = require("../models/subscription");
+const ErrorService = require("../middleware/error/errorService");
 
 async function getSubscriptionsByUser(userId) {
+  const result = await User.query().findById(userId);
+  if (!result) {
+    throw ErrorService.errorThrow(404);
+  }
   return Subscription.query().select({
     where: {
       subscriber: userId,
@@ -9,6 +16,10 @@ async function getSubscriptionsByUser(userId) {
 }
 
 async function getSubscribersByAuthor(userId) {
+  const result = await User.query().findById(userId);
+  if (!result) {
+    throw ErrorService.errorThrow(404);
+  }
   return Subscription.query().select({
     where: {
       author: userId,
@@ -21,6 +32,10 @@ async function createSubscription(data) {
 }
 
 async function removeSubscription(id) {
+  const result = await Subscription.query().findById(id);
+  if (!result) {
+    throw ErrorService.errorThrow(404);
+  }
   return Subscription.query().findById(id).delete();
 }
